tools/requirements.txt
anthropic>=0.25.0
networkx>=3.2
watchdog>=4.0
python-dotenv>=1.0.0
slack-sdk>=3.27.0
